# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Metrics Todo API",
    "version": "1.0.0",
    "description": "HTTP API for process metrics and user-scoped todo management. Todo endpoints require a Bearer JWT token."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "category": { "type": "string", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed"],
            "default": "pending"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "default": "medium"
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "last_viewed": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/email-reminders/last-sent": {
      "get": {
        "summary": "Get last email status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Last email status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "last_email_sent": { "type": "string", "format": "date-time", "nullable": true },
                    "last_checked_at": { "type": "string", "format": "date-time", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/email-reminders/due-todos": {
      "get": {
        "summary": "Get user's due todos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Due todos list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "todos": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Todo" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/email-reminders/send-all": {
      "post": {
        "summary": "Send reminder emails to all users (requires admin role)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Emails sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "result": {
                      "type": "object",
                      "properties": {
                        "successCount": { "type": "integer" },
                        "failureCount": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}